assets: Add new star candidate
The primary purpose of this is to fix the fact that the old star's top was
cut off slightly due to the assumption of an iPad-sized screen.  However,
I also added a bit of dimensionality to it and adjusted the glow a bit.
We can see how it looks and can revert to the old style of desired.

This adds 54 pixels above the old star, so it should be draw 54 pixels
higher up.